+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 
  - Improve RAM Input by adding System Free/Used RAM and Mycodo Frontend RAM
  - Add API endpoints to export Mycodo Configuration and Influxdb Database
+ - Add Input: System Uptime
 
 ### Miscellaneous